refactor: extract cache engine from `getVp`
