+++ src/methods.ts
@@ -1,7 +1,7 @@
 import snapshot from '@snapshot-labs/strategies';
-import redis from './redis';
-import { sha256, getCurrentBlockNum } from './utils';
 import disabled from './disabled.json';
+import { sha256, getCurrentBlockNum } from './utils';
+import { cachedVp } from './helpers/cache';
 
 interface GetVpRequestParams {
   address: string;
@@ -24,6 +24,9 @@ interface ValidateRequestParams {
 const disableCachingForSpaces = ['magicappstore.eth', 'moonbeam-foundation.eth'];
 
 export async function getVp(params: GetVpRequestParams) {
+  if (['1319'].includes(params.network) || disabled.includes(params.space))
+    throw 'something wrong with the strategies';
+
   if (typeof params.snapshot !== 'number') params.snapshot = 'latest';
 
   if (params.snapshot !== 'latest') {
@@ -31,41 +34,20 @@ export async function getVp(params: GetVpRequestParams) {
     params.snapshot = currentBlockNum < params.snapshot ? 'latest' : params.snapshot;
   }
 
-  const key = sha256(JSON.stringify(params));
-  const useCache =
-    redis && params.snapshot !== 'latest' && !disableCachingForSpaces.includes(params.space);
-  if (useCache) {
-    const cache = await redis.hGetAll(`vp:${key}`);
-
-    if (cache?.vp_state) {
-      cache.vp = parseFloat(cache.vp);
-
-      cache.vp_by_strategy = JSON.parse(cache.vp_by_strategy);
-      return { result: cache, cache: true };
-    }
-  }
-
-  if (['1319'].includes(params.network) || disabled.includes(params.space))
-    throw 'something wrong with the strategies';
-
-  const result = await snapshot.utils.getVp(
-    params.address,
-    params.network,
-    params.strategies,
-    params.snapshot,
-    params.space,
-    params.delegation
+  return await cachedVp(
+    sha256(JSON.stringify(params)),
+    async () => {
+      return await snapshot.utils.getVp(
+        params.address,
+        params.network,
+        params.strategies,
+        params.snapshot,
+        params.space,
+        params.delegation
+      );
+    },
+    params.snapshot !== 'latest' && !disableCachingForSpaces.includes(params.space)
   );
-
-  if (useCache && result.vp_state === 'final') {
-    const multi = redis.multi();
-    multi.hSet(`vp:${key}`, 'vp', result.vp);
-    multi.hSet(`vp:${key}`, 'vp_by_strategy', JSON.stringify(result.vp_by_strategy));
-    multi.hSet(`vp:${key}`, 'vp_state', result.vp_state);
-    multi.exec();
-  }
-
-  return { result, cache: false };
 }
 
 export async function validate(params: ValidateRequestParams) {
